+++ packages/core/src/constant.ts
@@ -1,26 +1,3 @@
-import type { TextRenderOption } from './utils/textRender'
-
-const TopicStyle = {
-  maxWidth: 240,
-  get minHeight() {
-    return this.rootTopicFontSize * this.lineHeight + this.padding * 2
-  },
-  padding: 8,
-  margin: 80,
-  borderWidth: 2,
-  rootTopicFontSize: 18,
-  fontFamily: `"Microsoft Yahei", "PingFang SC"`,
-  lineHeight: 1.2,
-}
-
-const TopicTextRenderOptions: Omit<TextRenderOption, 'style'> = {
-  box: {
-    width: TopicStyle.maxWidth,
-    height: 10_000,
-  },
-  padding: TopicStyle.padding,
-}
-
 enum EDITOR_MODE {
   none,
   drag,
@@ -33,4 +10,15 @@ enum ViewType {
   link,
 }
 
-export { EDITOR_MODE, TopicStyle, ViewType, TopicTextRenderOptions }
+enum Color {
+  white = '#fff',
+  black = '#141414',
+  red = '#F44336',
+}
+
+enum FontWeight {
+  Regular = '400',
+  Medium = '500',
+}
+
+export { EDITOR_MODE, ViewType, Color, FontWeight }