+++ packages/core/src/layout/index.ts
@@ -1,17 +1,20 @@
-import { LayoutType } from '../interface/layout'
+import { LayoutOption } from '../interface/layout'
 import { LayoutTopic, TopicData } from '../interface/topic'
 import { mindmap } from './mindmap'
 import { getCanvasSize } from './shared'
-import { structure } from './tree'
+import { structure } from './structure'
 
-function doLayout(root: TopicData, layout: LayoutType) {
+const canvasPadding = 10
+
+function doLayout(root: TopicData, options: LayoutOption) {
+  const { layout, theme } = options
   let layoutRoot: LayoutTopic
   switch (layout) {
     case 'mindmap':
-      layoutRoot = mindmap(root)
+      layoutRoot = mindmap(root, options)
       break
     case 'structure':
-      layoutRoot = structure(root)
+      layoutRoot = structure(root, options)
       break
     default:
       throw new Error('invalid layout')
@@ -19,10 +22,11 @@ function doLayout(root: TopicData, layout: LayoutType) {
 
   layoutRoot.descendants().forEach((node) => {
     // add canvas padding
-    node.x += 10
+    node.x += canvasPadding
+    node.y += canvasPadding
   })
 
-  const [canvasWidth, canvasHeight] = getCanvasSize(layoutRoot)
+  const [canvasWidth, canvasHeight] = getCanvasSize(theme, layoutRoot)
   return {
     layoutRoot,
     canvasWidth,