+++ packages/core/src/view/Topic.tsx
@@ -1,12 +1,7 @@
 import { useContext } from 'preact/hooks'
 import { ThemeContext } from '../context/theme'
-import {
-  EDITOR_MODE,
-  TopicStyle,
-  TopicTextRenderOptions,
-  ViewType,
-} from '../constant'
-import { getTopicTextStyle } from '../layout/shared'
+import { EDITOR_MODE, ViewType } from '../constant'
+import { createTopicTextRenderBox, getTopicTextStyle } from '../layout/shared'
 import { ViewModel } from '../viewModel'
 import { LayoutTopic } from '../interface/topic'
 import { renderText } from '../utils/textRender'
@@ -18,32 +13,32 @@ type TopicProps = {
 
 const Topic = (props: TopicProps) => {
   const viewModel = ViewModel.useContainer()
-  const $theme = useContext(ThemeContext)
+  const theme = useContext(ThemeContext)
+  const {
+    topic: { borderWidth, borderColor },
+  } = theme
   const { node } = props
   const {
     data: { title, id },
     x,
     y,
-    depth,
     size,
   } = node
   const { mode, selection } = viewModel
   const isSelected = id === selection
   const isEditing = isSelected && mode === EDITOR_MODE.edit
-  const isMainTopic = depth <= 1
   const [width, height] = size
 
   const outline = isSelected
     ? {
-        stroke: $theme.mainColor,
-        strokeWidth: TopicStyle.borderWidth,
+        stroke: borderColor,
+        strokeWidth: borderWidth,
       }
     : {}
-  const background = isMainTopic || isEditing ? '#fff' : 'transparent'
 
-  const textStyle = getTopicTextStyle(node)
+  const textStyle = getTopicTextStyle(theme, node)
   const { lines } = renderText(title, {
-    ...TopicTextRenderOptions,
+    ...createTopicTextRenderBox(theme, node),
     style: textStyle,
   })
 
@@ -58,11 +53,12 @@ const Topic = (props: TopicProps) => {
       <rect
         width={width}
         height={height}
-        fill={background}
-        radius={5}
+        fill={textStyle.background}
+        rx={textStyle.borderRadius}
+        ry={textStyle.borderRadius}
         {...outline}
       ></rect>
-      <text style={textStyle}>
+      <text style={textStyle} fill={textStyle.color}>
         {lines.map((line) => (
           <tspan x={line.x} y={line.y}>
             {line.text}